+++ src/GammonX/GammonX.Mars.NN.Tests/Services/FevgaFeatureEvalServiceTests.cs
@@ -55,7 +55,8 @@ namespace GammonX.Mars.NN.Tests.Services
                 Board = boardContract,
                 Modus = GameModus.Fevga,
                 Rolls = [roll1, roll2],
-                IsWhite = true
+                IsWhite = true,
+                BotLevel = BotLevel.Hard
             };
             var resultWhite = evalService.EvalMoveSequence(requestWhite, EvalWeights.FevgaCheapContactWeights, EvalWeights.FevgaContactWeights, EvalWeights.RaceWeights, 20);
             Assert.NotNull(resultWhite);
@@ -65,7 +66,8 @@ namespace GammonX.Mars.NN.Tests.Services
                 Board = boardContract,
                 Modus = GameModus.Fevga,
                 Rolls = [roll1, roll2],
-                IsWhite = false
+                IsWhite = false,
+                BotLevel = BotLevel.Hard
             };
             var resultBlack = evalService.EvalMoveSequence(requestBlack, EvalWeights.FevgaCheapContactWeights, EvalWeights.FevgaContactWeights, EvalWeights.RaceWeights, 20);
             Assert.NotNull(resultBlack);
@@ -98,7 +100,8 @@ namespace GammonX.Mars.NN.Tests.Services
                 Board = boardContract,
                 Modus = GameModus.Fevga,
                 Rolls = [roll1, roll2, roll1, roll2],
-                IsWhite = true
+                IsWhite = true,
+                BotLevel = BotLevel.Hard
             };
             var resultWhite = evalService.EvalMoveSequence(requestWhite, EvalWeights.FevgaCheapContactWeights, EvalWeights.FevgaContactWeights, EvalWeights.RaceWeights, 20);
             Assert.NotNull(resultWhite);
@@ -108,7 +111,8 @@ namespace GammonX.Mars.NN.Tests.Services
                 Board = boardContract,
                 Modus = GameModus.Fevga,
                 Rolls = [roll1, roll2, roll1, roll2],
-                IsWhite = false
+                IsWhite = false,
+                BotLevel = BotLevel.Hard
             };
             var resultBlack = evalService.EvalMoveSequence(requestBlack, EvalWeights.FevgaCheapContactWeights, EvalWeights.FevgaContactWeights, EvalWeights.RaceWeights, 20);
             Assert.NotNull(resultBlack);
@@ -139,7 +143,8 @@ namespace GammonX.Mars.NN.Tests.Services
                 Board = boardContract,
                 Modus = GameModus.Fevga,
                 Rolls = [roll1, roll2, roll1, roll2],
-                IsWhite = true
+                IsWhite = true,
+                BotLevel = BotLevel.Hard
             };
             var resultWhite = evalService.EvalMoveSequence(requestWhite, EvalWeights.FevgaCheapContactWeights, EvalWeights.FevgaContactWeights, EvalWeights.RaceWeights, 20);
             Assert.NotNull(resultWhite);
@@ -149,7 +154,8 @@ namespace GammonX.Mars.NN.Tests.Services
                 Board = boardContract,
                 Modus = GameModus.Fevga,
                 Rolls = [roll1, roll2, roll1, roll2],
-                IsWhite = false
+                IsWhite = false,
+                BotLevel = BotLevel.Hard
             };
             var resultBlack = evalService.EvalMoveSequence(requestBlack, EvalWeights.FevgaCheapContactWeights, EvalWeights.FevgaContactWeights, EvalWeights.RaceWeights, 20);
             Assert.NotNull(resultBlack);
@@ -177,7 +183,8 @@ namespace GammonX.Mars.NN.Tests.Services
                 Board = boardContract,
                 Modus = GameModus.Fevga,
                 Rolls = [roll1, roll2, roll1, roll2],
-                IsWhite = true
+                IsWhite = true,
+                    BotLevel = BotLevel.Hard
             };
             var resultWhite = evalService.EvalMoveSequence(requestWhite, EvalWeights.FevgaCheapContactWeights, EvalWeights.FevgaContactWeights, EvalWeights.RaceWeights, 20);
             Assert.NotNull(resultWhite);
@@ -230,7 +237,8 @@ namespace GammonX.Mars.NN.Tests.Services
                 IsWhite = false,
                 MatchLength = 2,
                 PointsAwayOpp = 1,
-                PointsAwayPlayer = 1
+                PointsAwayPlayer = 1,
+                BotLevel = BotLevel.Hard
             };
 
             Assert.Throws<InvalidOperationException>(() => evalService.EvalCube(request));