+++ src/GammonX/GammonX.Mars.Server/Program.cs
@@ -125,6 +125,3 @@ app.UseAuthorization();
 app.MapHealthChecks("/health");
 app.MapControllers();
 app.Run();
-
-// we require this for web application factory tests
-public partial class Program { }