+++ src/GammonX/GammonX.Server/Models/matchSession/IMatchSessionModel.cs
@@ -2,6 +2,8 @@
 
 using GammonX.Server.Contracts;
 
+using MatchType = GammonX.Models.Enums.MatchType;
+
 namespace GammonX.Server.Models
 {
 	/// <summary>
@@ -32,7 +34,12 @@ namespace GammonX.Server.Models
 		/// <summary>
 		/// Gets the type of the match determining the winning condition.
 		/// </summary>
-		GammonX.Models.Enums.MatchType Type { get; }
+		MatchType Type { get; }
+
+        /// <summary>
+        /// Gets the bot level if this match session is played against a bot. Otherwise, <see cref="BotLevel.Unknown"/>.
+        /// </summary>
+        BotLevel BotLevel { get; }
 
 		/// <summary>
 		/// Gets the player 1.